app/requirements/staging.txt
Django==3.0.5

psycopg2-binary==2.8.5

djangorestframework==3.11.0
requests==2.23.0
django-choices==1.7.1
django-extensions==2.2.9
ipython==7.11.1
django-filter==2.2.0
factory-boy==2.12.0
